src/z_stuff/requirements.txt
altair
boto3
dash
dateparser
docker
fastapi
flask
google-cloud-bigquery
google-cloud-core
google-cloud-sqlcommenter
google-cloud-storage
matplotlib
missingno
nltk
numpy
openpyxl
pandas
plotly
polars
psycopg2-binary
pyarrow
pyjanitor
scikit-learn
scipy
seaborn
spacy
SQLAlchemy
streamlit
textblob
transformers
uvicorn
xgboost

